Add files via upload
 Number Guessing Game
This number guessing game is an easy project built on Java where the player has to guess a number given in between a range. 
If the guessed number is right, the player wins else, loses. It also has the concept of limited attempts where the player has 
to guess the number within the limited attempts given. 

Abstract: The UI has an input value option where the player has to enter the guessed value, it also displays the time remaining 
to guess. After completing the limits given, if the guessed number is right, the player wins else loses. The range between the 
number can be from 1 to 100 or 1 to 1000. Also, if the number you’ve guessed is high or low to the actual value, the application 
sends you an alert “Too High” or “Too Low”. After the limited attempt is completed, the actual value is revealed.